% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{combined_data}
\alias{combined_data}
\title{Conference Data}
\format{
## `combined_data`
A data frame with 204 rows and 8 columns:
\describe{
  \item{talk_title}{Title}
  \item{page_value}{Value for page panel}
  \item{session_type}{Type}
  \item{session_topics}{Topics}
  \item{speaker_name}{Speaker}
  \item{track}{Track}
  \item{description}{Description}
  \item{yt_url}{YouTube Video URL}
}
}
\source{
<https://raw.githubusercontent.com/rfordatascience/tidytuesday/main/data/2025/2025-01-14/conf2024.csv>
}
\usage{
combined_data
}
\description{
Conference Data
}
\keyword{datasets}
