% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_talks.R
\name{create_speaker_conditions}
\alias{create_speaker_conditions}
\title{Create search term conditions}
\usage{
create_speaker_conditions(speaker)
}
\arguments{
\item{speaker}{speaker name}
}
\value{
string that contains javascript code for search term conditions
}
\description{
Create search term conditions
}
