% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_talks.R
\name{create_talk_page}
\alias{create_talk_page}
\title{Create a talk page}
\usage{
create_talk_page(page_data)
}
\arguments{
\item{page_data}{Data for a talk page}
}
\value{
div
}
\description{
Create a talk page
}
