% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_talks.R
\name{create_topic_conditions}
\alias{create_topic_conditions}
\title{Create session topic selection conditions}
\usage{
create_topic_conditions(topics)
}
\arguments{
\item{topics}{Vector of session topics}
}
\value{
string that contains javascript code for session topic conditions
}
\description{
Create session topic selection conditions
}
