% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_talks.R
\name{create_talk_card}
\alias{create_talk_card}
\title{Create talk card}
\usage{
create_talk_card(talk_title, speaker, topics, page_value)
}
\arguments{
\item{talk_title}{talk title}

\item{speaker}{speaker name}

\item{topics}{session topics}

\item{page_value}{page ID}
}
\value{
conditional panel
}
\description{
Create talk card
}
